# perf(mapper): avoid O(N^2) algorithm