@lazytanuki 2022-06-05
Oh nice ! I didn't realize using the vimscript bridge would have this much of an impact, thanks for the work and explanation !